Fix: using ulong for the neibslistSize
On devices with a lot of memory, the number of allocated particles can be
big enough to make neibslistSize overflow. The neibslistSize is at least
100 times bigger than the other sizes (numParts*maxneibsnum*sizeof...)
so the same should not happen for the other buffers.
TODO: use size_t instead.

+++ src/GPUWorker.cc
@@ -771,7 +771,8 @@ size_t GPUWorker::allocateDeviceBuffers() {
 	const uint infoSize = sizeof(particleinfo) * m_numAllocatedParticles;
 	const uint intSize = sizeof(uint) * m_numAllocatedParticles;
 	const uint uintCellsSize = sizeof(uint) * m_nGridCells;
-	const uint neibslistSize = sizeof(uint) * m_simparams->maxneibsnum*(m_numAllocatedParticles/NEIBINDEX_INTERLEAVE + 1)*NEIBINDEX_INTERLEAVE;
+	// ulong for the neibslistSize since it is at least 2 orders of magnitude bigger than the other sizes and could overflow
+	const ulong neibslistSize = sizeof(uint) * m_simparams->maxneibsnum*(m_numAllocatedParticles/NEIBINDEX_INTERLEAVE + 1)*NEIBINDEX_INTERLEAVE;
 	const uint hashSize = sizeof(hashKey) * m_numAllocatedParticles;
 	const uint segmentsSize = sizeof(uint) * 4; // 4 = types of cells
 	//const uint neibslistSize = sizeof(uint) * 128 * m_numAlocatedParticles;
